feat(backfill): add eth supply backfill functionality
Introduced a new backfill process for the eth supply table, allowing
for the retrieval and storage of missing entries based on prerequisite
data. This includes new SQL queries and a command in the CLI to
initiate the backfill, enhancing data integrity in the beacon chain.
New queries for counting and fetching missing slots have also been
added to support this functionality.